requirements.txt
httpx_socks
httpx
openai
aiogram
python-dotenv